Keeping fixed literal reasons alive:
By default, all trail literals of level 0 have a nullptr reason. These literals are considered as fixed and inactive.
In this patch, fixed literals which had a driving clause as a reason, maintained it. In addition, these literals are
reactivated in undo_trail_core method so it would be able to propagate them again internally.